pyproject.toml
[project]
name = "constructor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "openai>=2.11.0",
    "python-dotenv>=1.2.1",
]
